+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## v0.17.4
 
 FEATURES: